+++ python/pyarrow/tests/parquet/test_datetime.py
@@ -18,9 +18,8 @@
 import datetime
 
 import numpy as np
-import pytest
-
 import pyarrow as pa
+import pytest
 from pyarrow.tests.parquet.common import _check_roundtrip
 
 try:
@@ -33,7 +32,6 @@ except ImportError:
 try:
     import pandas as pd
     import pandas.testing as tm
-
     from pyarrow.tests.parquet.common import _roundtrip_pandas_dataframe
 except ImportError:
     pd = tm = None
